Further tweaks

+++ ffs/tests/ffs_file_test.c
@@ -19,26 +19,26 @@
 static int verbose = 0;
 static char *test_only = NULL;
 
-int
-size_func_sizeof(FFSFile iofile, int size)
+size_t
+size_func_sizeof(FFSFile iofile, size_t size)
 {
     return size;
 }
 
-int
-size_func_next_size(FFSFile iofile, int size)
+size_t
+size_func_next_size(FFSFile iofile, size_t size)
 {
     return FFSfile_next_decode_length(iofile);
 }
 
 int
-read_func_no_buffer(FFSFile iofile, void *data, int size)
+read_func_no_buffer(FFSFile iofile, void *data, size_t size)
 {
     return FFSread(iofile, data);
 }
 
 int
-read_func_buffer(FFSFile iofile, void *data, int size)
+read_func_buffer(FFSFile iofile, void *data, size_t size)
 {
     FFSBuffer b = create_fixed_FFSBuffer(data, size);
     int ret = FFSread_to_buffer(iofile, b, NULL);
@@ -102,7 +102,7 @@ set_targets(FFSContext context)
 }
 
 void
-do_test(char *input_file, int (*size_func) (FFSFile, int), int (*read_func) (FFSFile iofile, void *data, int size))
+do_test(char *input_file, size_t (*size_func) (FFSFile, size_t), int (*read_func) (FFSFile iofile, void *data, size_t size))
 {
     FFSFile iofile;
     int finished = 0;
@@ -172,7 +172,7 @@ do_test(char *input_file, int (*size_func) (FFSFile, int), int (*read_func) (FFS
 		    exit(1);
 		}
 	    } else if (FFSnext_type_handle(iofile) == second_rec_ioformat) {
-		int size = size_func(iofile, sizeof(second_rec));
+		size_t size = size_func(iofile, sizeof(second_rec));
 		second_rec *read_data = malloc(size);
 		memset(read_data, 0, size);
 		if (!read_func(iofile, read_data, size))
@@ -183,7 +183,7 @@ do_test(char *input_file, int (*size_func) (FFSFile, int), int (*read_func) (FFS
 		}
 		free(read_data);
 	    } else if (FFSnext_type_handle(iofile) == third_rec_ioformat) {
-		int size = size_func(iofile, sizeof(third_rec));
+		size_t size = size_func(iofile, sizeof(third_rec));
 		third_rec *read_data = malloc(size);
 		memset(read_data, 0, size);
 		if (!read_func(iofile, read_data, size))
@@ -194,7 +194,7 @@ do_test(char *input_file, int (*size_func) (FFSFile, int), int (*read_func) (FFS
 		}
 		free(read_data);
 	    } else if (FFSnext_type_handle(iofile) == fourth_rec_ioformat) {
-		int size = size_func(iofile, sizeof(fourth_rec));
+		size_t size = size_func(iofile, sizeof(fourth_rec));
 		fourth_rec *read_data = malloc(size);
 		memset(read_data, 0, size);
 		if (!read_func(iofile, read_data, size))
@@ -210,7 +210,7 @@ do_test(char *input_file, int (*size_func) (FFSFile, int), int (*read_func) (FFS
 		printf("Emb Rec failure\n");
 		FFSread(iofile, NULL);
 	    } else if (FFSnext_type_handle(iofile) == fifth_rec_ioformat) {
-		int size = size_func(iofile, sizeof(fifth_rec));
+		size_t size = size_func(iofile, sizeof(fifth_rec));
 		fifth_rec *read_data = malloc(size);
 		memset(read_data, 0, size);
 		if (!read_func(iofile, read_data, size))
@@ -222,7 +222,7 @@ do_test(char *input_file, int (*size_func) (FFSFile, int), int (*read_func) (FFS
 		free(read_data);
 		fifth_rec_count++;
 	    } else if (FFSnext_type_handle(iofile) == sixth_rec_ioformat) {
-		int size = size_func(iofile, sizeof(sixth_rec));
+		size_t size = size_func(iofile, sizeof(sixth_rec));
 		sixth_rec *read_data = malloc(size);
 		memset(read_data, 0, size);
 		if (!read_func(iofile, read_data, size))
@@ -233,7 +233,7 @@ do_test(char *input_file, int (*size_func) (FFSFile, int), int (*read_func) (FFS
 		}
 		free(read_data);
 	    } else if (FFSnext_type_handle(iofile) == nested_rec_ioformat) {
-		int size = size_func(iofile, sizeof(nested_rec));
+		size_t size = size_func(iofile, sizeof(nested_rec));
 		nested_rec *read_data = malloc(size);
 		memset(read_data, 0, size);
 		if (!read_func(iofile, read_data, size))
@@ -244,7 +244,7 @@ do_test(char *input_file, int (*size_func) (FFSFile, int), int (*read_func) (FFS
 		}
 		free(read_data);
 	    } else if (FFSnext_type_handle(iofile) == later_rec_ioformat) {
-		int size = size_func(iofile, sizeof(later_rec));
+		size_t size = size_func(iofile, sizeof(later_rec));
 		later_rec *read_data = malloc(size);
 		memset(read_data, 0, size);
 		if (!read_func(iofile, read_data, size))
@@ -255,7 +255,7 @@ do_test(char *input_file, int (*size_func) (FFSFile, int), int (*read_func) (FFS
 		}
 		free(read_data);
 	    } else if (FFSnext_type_handle(iofile) == ninth_rec_ioformat) {
-		int size = size_func(iofile, sizeof(ninth_rec));
+		size_t size = size_func(iofile, sizeof(ninth_rec));
 		ninth_rec *read_data = malloc(size);
 		memset(read_data, 0, size);
 		if (!read_func(iofile, read_data, size))
@@ -266,7 +266,7 @@ do_test(char *input_file, int (*size_func) (FFSFile, int), int (*read_func) (FFS
 		}
 		free(read_data);
 	    } else if (FFSnext_type_handle(iofile) == string_array_ioformat) {
-		int size = size_func(iofile, sizeof(string_array_rec));
+		size_t size = size_func(iofile, sizeof(string_array_rec));
 		string_array_rec *read_data = malloc(size);
 		memset(read_data, 0, size);
 		if (!read_func(iofile, read_data, size))
