+++ ffs/tests/ffs_index_test.c
@@ -19,14 +19,14 @@
 static int verbose = 0;
 static char *test_only = NULL;
 
-int
-size_func_sizeof(FFSFile iofile, int size)
+size_t
+size_func_sizeof(FFSFile iofile, size_t size)
 {
     return size;
 }
 
-int
-size_func_next_size(FFSFile iofile, int size)
+size_t
+size_func_next_size(FFSFile iofile, size_t size)
 {
     return FFSfile_next_decode_length(iofile);
 }
@@ -34,13 +34,13 @@ size_func_next_size(FFSFile iofile, int size)
 attr_list last_attrs = NULL;
 
 int
-read_func_no_buffer(FFSFile iofile, void *data, int size)
+read_func_no_buffer(FFSFile iofile, void *data, size_t size)
 {
     return FFSread_attr(iofile, data, &last_attrs);
 }
 
 int
-read_func_buffer(FFSFile iofile, void *data, int size)
+read_func_buffer(FFSFile iofile, void *data, size_t size)
 {
     FFSBuffer b = create_fixed_FFSBuffer(data, size);
     int ret = FFSread_to_buffer(iofile, b, NULL);
@@ -105,7 +105,7 @@ set_targets(FFSContext context)
 }
 
 void
-do_test(char *input_file, int (*size_func) (FFSFile, int), int (*read_func) (FFSFile iofile, void *data, int size))
+do_test(char *input_file, size_t (*size_func) (FFSFile, size_t), int (*read_func) (FFSFile iofile, void *data, size_t size))
 {
     FFSFile iofile;
     int item_count = 0;
@@ -203,7 +203,7 @@ do_test(char *input_file, int (*size_func) (FFSFile, int), int (*read_func) (FFS
 		    exit(1);
 		}
 	    } else if (FFSnext_type_handle(iofile) == second_rec_ioformat) {
-		int size = size_func(iofile, sizeof(second_rec));
+		size_t size = size_func(iofile, sizeof(second_rec));
 		second_rec *read_data = malloc(size);
 		int iter_attr = -1;
 		memset(read_data, 0, size);
@@ -221,7 +221,7 @@ do_test(char *input_file, int (*size_func) (FFSFile, int), int (*read_func) (FFS
 		second_rec_count++;
 		free(read_data);
 	    } else if (FFSnext_type_handle(iofile) == third_rec_ioformat) {
-		int size = size_func(iofile, sizeof(third_rec));
+		size_t size = size_func(iofile, sizeof(third_rec));
 		third_rec *read_data = malloc(size);
 		int iter_attr = -1;
 		memset(read_data, 0, size);
@@ -236,7 +236,7 @@ do_test(char *input_file, int (*size_func) (FFSFile, int), int (*read_func) (FFS
 		free(read_data);
 		third_rec_count++;
 	    } else if (FFSnext_type_handle(iofile) == fourth_rec_ioformat) {
-		int size = size_func(iofile, sizeof(fourth_rec));
+		size_t size = size_func(iofile, sizeof(fourth_rec));
 		fourth_rec *read_data = malloc(size);
 		memset(read_data, 0, size);
 		if (!read_func(iofile, read_data, size))
@@ -252,7 +252,7 @@ do_test(char *input_file, int (*size_func) (FFSFile, int), int (*read_func) (FFS
 		printf("Emb Rec failure\n");
 		FFSread(iofile, NULL);
 	    } else if (FFSnext_type_handle(iofile) == fifth_rec_ioformat) {
-		int size = size_func(iofile, sizeof(fifth_rec));
+		size_t size = size_func(iofile, sizeof(fifth_rec));
 		fifth_rec *read_data = malloc(size);
 		int iter_attr = -1;
 		memset(read_data, 0, size);
@@ -266,7 +266,7 @@ do_test(char *input_file, int (*size_func) (FFSFile, int), int (*read_func) (FFS
 		free(read_data);
 		fifth_rec_count++;
 	    } else if (FFSnext_type_handle(iofile) == sixth_rec_ioformat) {
-		int size = size_func(iofile, sizeof(sixth_rec));
+		size_t size = size_func(iofile, sizeof(sixth_rec));
 		sixth_rec *read_data = malloc(size);
 		memset(read_data, 0, size);
 		if (!read_func(iofile, read_data, size))
@@ -277,7 +277,7 @@ do_test(char *input_file, int (*size_func) (FFSFile, int), int (*read_func) (FFS
 		}
 		free(read_data);
 	    } else if (FFSnext_type_handle(iofile) == nested_rec_ioformat) {
-		int size = size_func(iofile, sizeof(nested_rec));
+		size_t size = size_func(iofile, sizeof(nested_rec));
 		nested_rec *read_data = malloc(size);
 		memset(read_data, 0, size);
 		if (!read_func(iofile, read_data, size))
@@ -289,7 +289,7 @@ do_test(char *input_file, int (*size_func) (FFSFile, int), int (*read_func) (FFS
 		}
 		free(read_data);
 	    } else if (FFSnext_type_handle(iofile) == later_rec_ioformat) {
-		int size = size_func(iofile, sizeof(later_rec));
+		size_t size = size_func(iofile, sizeof(later_rec));
 		later_rec *read_data = malloc(size);
 		memset(read_data, 0, size);
 		if (!read_func(iofile, read_data, size))
@@ -300,7 +300,7 @@ do_test(char *input_file, int (*size_func) (FFSFile, int), int (*read_func) (FFS
 		}
 		free(read_data);
 	    } else if (FFSnext_type_handle(iofile) == ninth_rec_ioformat) {
-		int size = size_func(iofile, sizeof(ninth_rec));
+		size_t size = size_func(iofile, sizeof(ninth_rec));
 		ninth_rec *read_data = malloc(size);
 		memset(read_data, 0, size);
 		if (!read_func(iofile, read_data, size))
@@ -311,7 +311,7 @@ do_test(char *input_file, int (*size_func) (FFSFile, int), int (*read_func) (FFS
 		}
 		free(read_data);
 	    } else if (FFSnext_type_handle(iofile) == string_array_ioformat) {
-		int size = size_func(iofile, sizeof(string_array_rec));
+		size_t size = size_func(iofile, sizeof(string_array_rec));
 		string_array_rec *read_data = malloc(size);
 		memset(read_data, 0, size);
 		if (!read_func(iofile, read_data, size))
